refactor: remove logic from index file
