+++ packages/core/src/modules/dids/domain/service/index.ts
@@ -1,42 +1,7 @@
-import type { ClassConstructor } from 'class-transformer'
-
-import { Transform, plainToInstance } from 'class-transformer'
-
 import { DidCommService } from './DidCommService'
 import { DidCommV2Service } from './DidCommV2Service'
 import { DidDocumentService } from './DidDocumentService'
 import { IndyAgentService } from './IndyAgentService'
+import { ServiceTransformer, serviceTypes } from './ServiceTransformer'
 
-export const serviceTypes: { [key: string]: unknown | undefined } = {
-  [IndyAgentService.type]: IndyAgentService,
-  [DidCommService.type]: DidCommService,
-  [DidCommV2Service.type]: DidCommV2Service,
-}
-
-/**
- * Decorator that transforms service json to corresponding class instances. See {@link serviceTypes}
- *
- * @example
- * class Example {
- *   ServiceTransformer()
- *   private service: Service
- * }
- */
-export function ServiceTransformer() {
-  return Transform(
-    ({ value }: { value: { type: string }[] }) => {
-      return value.map((serviceJson) => {
-        const serviceClass = (serviceTypes[serviceJson.type] ??
-          DidDocumentService) as ClassConstructor<DidDocumentService>
-        const service = plainToInstance<DidDocumentService, unknown>(serviceClass, serviceJson)
-
-        return service
-      })
-    },
-    {
-      toClassOnly: true,
-    }
-  )
-}
-
-export { IndyAgentService, DidCommService, DidDocumentService }
+export { IndyAgentService, DidCommService, DidDocumentService, DidCommV2Service, ServiceTransformer, serviceTypes }
